SISTEMA DE VACINACAO - FINAL/requirements.txt
# NASST Digital - Dependências
# Versão 1.1

# Core
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Banco de Dados
sqlite3  # Built-in, não precisa instalar

# Visualização
plotly>=5.14.0

# Excel/CSV
openpyxl>=3.1.0
xlrd>=2.0.1

# PDF
pdfplumber>=0.10.0
reportlab>=4.0.0
fpdf>=1.7.2

# Utilitários
python-dotenv>=1.0.0
schedule>=1.2.0

# Testes e Desenvolvimento
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
coverage==7.2.7
black==23.3.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3